requirements.txt
Flask
requests
loguru
azure.storage.blob
azure.core